Re-structued a bit
