+++ kernel/Cargo.toml
@@ -21,6 +21,7 @@ x86_64c = { package = "x86_64", version = "0.14.2" }
 uart_16550 = "0.3.2"
 acpic = { package = "acpi", version = "5.2.0" }
 talc = "4.4.2"
+nodit = "0.9.2"
 
 [[test]]
 name = "should_panic"
