+++ kernel/src/arch/x86_64/acpi/mod.rs
@@ -1,0 +1,1 @@
+pub mod handler;
