+++ kernel/src/arch/x86_64/mod.rs
@@ -16,6 +16,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
+pub mod acpi;
 pub mod clock;
 pub mod debug;
 pub mod interrupts;
